backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2",
    "django-allauth>=65.7.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dotenv>=0.9.9",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
